# docs/swagger.yaml
basePath: /x_ray
definitions:
...
  dto.CreateReviewReq:
    properties:
      benefit:
        description: rating out of 5
        type: integer
      business_outlook:
        type: string
      ceo_approval:
        type: boolean
      company_id:
        type: string
      compensation:
        description: rating out of 5
        type: integer
      con:
        type: string
      pro:
        type: string
      recommend:
        type: boolean
      star_rating:
        description: overall rating out of 5
        type: integer
      user_id:
        type: string
      work_life_balance:
        description: rating out of 5
        type: integer
    type: object
...
  dto.ReviewRes:
    properties:
      benefit:
        description: rating out of 5
        type: integer
      business_outlook:
        type: string
      ceo_approval:
        type: boolean
      company_id:
        type: string
      compensation:
        description: rating out of 5
        type: integer
      con:
        type: string
      id:
        type: string
      pro:
        type: string
      recommend:
        type: boolean
      star_rating:
        description: overall rating out of 5
        type: integer
      work_life_balance:
        description: rating out of 5
        type: integer
    type: object
...
  localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      time:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: this Backend is for x-ray, a company review platform. this backend/project
    was build with the intention to providing transparency to the job seekers, and
    also to give the company a chance to improve their work environment. this project
    is still in early stage, so expect some breaking changes in the future.
  title: X-ray API
  version: "1.0"
paths:
...
  /review:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of reviews with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ReviewRes'
                  type: array
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Reviews by Pagination
      tags:
      - Review
    post:
      consumes:
      - application/json
      description: Creates a new review for a company (requires authentication)
      parameters:
      - description: Review details
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: Review created successfully
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create Review
      tags:
      - Review
  /review/{review_id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific review
      parameters:
      - description: Review ID (UUID)
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReviewRes'
              type: object
        "400":
          description: Missing review_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Review by ID
      tags:
      - Review
  /review/company/{company_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews for a specific company with pagination
      parameters:
      - description: Company ID (UUID)
        in: path
        name: company_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ReviewRes'
                  type: array
              type: object
        "400":
          description: Missing company_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Reviews by Company ID
      tags:
      - Review
  /review/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews created by a specific user with pagination
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reviews fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ReviewRes'
                  type: array
              type: object
        "400":
          description: Missing user_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Reviews by User ID
      tags:
      - Review
...
swagger: "2.0"